--- JSON/author.json
@@ -1,10 +1,0 @@
-[
-  {
-    "first_name": "zdv",
-    "last_name": "zsdc"
-  },
-  {
-    "first_name": "zasdf",
-    "last_name": "sf"
-  }
-]
