+++ JSON/authors.json
@@ -1,0 +1,14 @@
+[
+  {
+    "first_name": "sfsdf",
+    "last_name": "sdfsdf"
+  },
+  {
+    "first_name": "dfdf",
+    "last_name": "sfsf"
+  },
+  {
+    "first_name": "dfsdf",
+    "last_name": "sfsdf"
+  }
+]
