--- JSON/game.json
@@ -1,12 +1,0 @@
-[
-  {
-    "name": "9",
-    "multiplayer": true,
-    "last_played_at": 0
-  },
-  {
-    "name": "asdf",
-    "multiplayer": true,
-    "last_played_at": 0
-  }
-]
