+++ JSON/games.json
@@ -1,0 +1,22 @@
+[
+  {
+    "name": "dfsdf",
+    "multiplayer": true,
+    "last_played_at": 0
+  },
+  {
+    "name": "sfsdf",
+    "multiplayer": false,
+    "last_played_at": 0
+  },
+  {
+    "name": "dfsdf",
+    "multiplayer": true,
+    "last_played_at": 0
+  },
+  {
+    "name": "fsfsdf",
+    "multiplayer": true,
+    "last_played_at": 0
+  }
+]
