+++ JSON/genre.json
@@ -4,5 +4,11 @@
   },
   {
     "name": "Hip Hip"
+  },
+  {
+    "name": "sfdff"
+  },
+  {
+    "name": "fsdfs"
   }
 ]
