+++ JSON/labels.json
@@ -6,5 +6,17 @@
   {
     "title": "fgjxk",
     "color": "xfyhk"
+  },
+  {
+    "title": "sfdfsdf",
+    "color": "dfsdf"
+  },
+  {
+    "title": "dfsdfsd",
+    "color": "sdfsd"
+  },
+  {
+    "title": "fsfsdf",
+    "color": "dfsdf"
   }
 ]
