+++ JSON/music_albums.json
@@ -10,5 +10,17 @@
     "publish_date": "2006",
     "on_spotify": false,
     "genre": "Hip Hip"
+  },
+  {
+    "name": "dfsdf",
+    "publish_date": "2000",
+    "on_spotify": true,
+    "genre": "sfdff"
+  },
+  {
+    "name": "dfsd",
+    "publish_date": "sfdf",
+    "on_spotify": true,
+    "genre": "fsdfs"
   }
 ]
