+++ class/book.rb
@@ -1,4 +1,4 @@
-require './item'
+require_relative './item'
 
 class Book < Item
   attr_accessor :publisher, :cover_state, :publish_date, :author, :title
