+++ class/genre.rb
