+++ class/item.rb
