+++ class/label.rb
