+++ main.rb
@@ -1,4 +1,4 @@
-require './app'
+require './class/app'
 
 def choose_options(input, app)
   case input
@@ -20,8 +20,8 @@ def choose_options(input, app)
     app.add_musicalbum
   when 9
     app.add_game
-  when 10
-    app.store_data
+  # when 10
+  #   app.store_data
   end
 end
