+++ preserve/data.rb
@@ -1,5 +1,5 @@
-require_relative './book'
-require_relative './app'
+require_relative '../class/book'
+require_relative '../class/app'
 
 # This part is to save book in json files
 def save_book(publisher, cover_state, publish_date, author, title)
