+++ spec/helper_spec.rb
@@ -1,0 +1,9 @@
+require_relative '../class/label'
+require_relative '../class/book'
+require_relative '../class/author'
+require_relative '../class/game'
+require_relative '../class/genre'
+require_relative '../class/music'
+require_relative '../preserve/game_author_preserve'
+require_relative '../preserve/save_music_genre'
+require_relative '../preserve/data'
